// src/app/Model/ftpmaster.ts
export class Ftpmaster {
    ftpid: string;
    ftpcode: string;
    ftpname: string;
    ftpurl: string;
    ftpusername: string;
    ftppassword: string;
    ftpport: string;
    ftptype: string;
    status: boolean;
    ftp835inboundfolder: string;
    ftp837inboundfolder: string;
    ftp835outboundfolder: string;
}